Revised proposal for hs.next and hs.last
This pull request implements a revised proposal for the hs.next and
hs.last operations that avoids some of the complexity of the original
specification -- e.g., the need to possibly compute over arbitrarily
many elements to determine how to advance hs.next so that it refers to
the smallest index of an invalid entry.

In this proposal we simply intialize hs.next to the first element and
advance on each subsequent call to extract(hs.next).

Other changes:

* Tweak wording of variable initialization to clarify when hs.next is
  initialized.

* Add handling to parser abstract machines.

* Rework MPLS parsing example.