Grammar, typo, punctuation, and P4 syntax fixes.
Some of the rewording suggestions are judgement calls, but to my
native English speaker's ears the proposed changes sound more natural.

For the change involving the paragraph containing ```Checksum16```,
Madoko apparently formats it differently if that text appears first on
an input line, vs. later in a line.  The formatted PDF output looks
more like what I would expect with the change I made there.

In the first sample P4 program, I ran it through a recent version of
the p4c compiler.  It complained about these lines:

    Checksum16() ck;

but gave no such complaints when the () was removed.  I did not dig
deep to figure out whether that was a problem with p4c.