Another Madoko markup fix
Only the first part of this change actually causes badness in the PDF
output.  The others do not make any visual change in the output, but
they do eliminate the last of the occurrences of ``` as the first
non-whitespace on a line, except for the definitions of P4Example,
etc.