New PSA.mdk Counter section
There isn't any fleshing out of the counter control plane API with
these changes.

There is significant fleshing out of the Counter and DirectCounter
extern definitions from the P4 program perspective, including a
reference implementation defined as pseudocode, and an example P4
program fragment demonstrating the use of both Counter and
DirectCounter.

Perhaps controversial changes:

Eliminate CounterType_t.  If you want a packet counter, just
instantiate a counter and only use the count() method that defaults to
adding 1 to the counter.  If you want a byte counter, use the count()
method that takes an 'add_value' parameter, and pass whatever you want
as the number of bytes in the packet.  If you want both, instantiate 2
counter externs, and update them both.

Enable more than one DirectCounter to be associated with a single
table.

Specify that the counter update operation should be saturating, not
wrap-around, including the rationale for why.