Change direction of some Ingress and Egress parameters from out to inout
We want these structs to have fields with defined values when the
Ingress and Egress control blocks begin execution.  This cannot be
done if they are defined as direction 'out', but can if they are
defined as direction 'inout'.

Also adds a missing type parameter <T> to one emit method definition
in psa.p4 that caused compilation errors.