PSA updates to address comments
Replace occurrences of "PSA program" with "P4 program".

Eliminate explicit mentions of "does not create a clone" in the
example packet flow.

Mention that a PSA implementation may implement a limit on the number
of times a single original packet can be recirculated, resubmitted, or
cloned from egress to egress.