PSA Add proposed method to associate metadata with packets
This commit only updates 2 of the example programs so that they
compile correctly, not all of them.  The others can be updated in a
later commit.  The changes to the others would be pretty much the same
as the changes made to psa-example-counters.p4 in this commit.

This includes some proposed updates to the section on cloning, as
well, but not yet for the resubmit and recirculate sections, which
also need updating.

The example program psa-example-bridged-metadata.p4 should become an
example somewhere in the document of doing all of these operations in
the same program, with metadata associated on all packet paths that
support it.