Remove unnecessary psa_parser_output_metadata_t struct
The only reason it was there was to carry the parser error out of the
ingress and egress parsers, but the PSA architecture can just
transport those parser errors to the ingress and egress control blocks
'somehow' by however the architecture does it, without making it an
'out' parameter of the parsers.

Note that the example architecture in the P4_16 language specification
does the same thing.