PSA Remove duplicate Madoko include tags
Madoko gives a warning message, I believe the condition in this
situation causing it is: if there are multiple files in INCLUDE
directives with the same name/id after BEGIN/END, then the warning
occurs.

The warning goes away when I delete these two lines.  PSA.mdk only
uses this name to INCLUDE a sample of code from the file
examples/psa-example-parser-checksum.p4, not
examples/psa-example-parser-error-handling.p4, so remove the
occurrence that we do not use or need.