Corrected syntax error in psa.p4
According to the P4-16 grammar,
the enum declarations do not end with a semicolon
(unless they are in fact typedef declarations, wich is not the case here).

So there was a syntax error in the psa.p4 file, for the
declaration of the enum MeterColor_t, line 453 :
there was an unexpected semicolon at the end of the enum declaration.