To disambiguate the grammar, have the keyValueList s to be non-empty
If expressionList and keyValueList can both be empty,
then the grammar is ambiguous, since an annotations with
parenthesis but no arguments inside could either correspond
to an annotation with an empty expressionList or an annnotation with
an empty keyValueList.

The change suggested by this commit matches the
actual current implementation of the P4 compiler.