TypeArg can be nonTypeName, because of the return type of externs
The return type of extern functions can be a specialized type
receiving a type variable (hence a nonTypeName) as a type argument.
For example:
`extern widget<T> createWidget<T, U>(U a, T b);`
(Example taken from `factory2.p4` in `p4-c/testdata/p4_16_samples/`)

For this reason, the non-terminal `typeArg` needs to have
`nonTypeName` as one of its variants.

But this introduces a conflict:
when the parser sees
`nonTypeName '<' nonTypeName`
it can't know if it is a `expression '<' expression` of a comparison
or a `expression '<' typeArg ...` of the begining of a function call.

This is why the non-terminal realTypeArg (corresponding to the old typeArg)
was introduced: to avoid that conflict.

This modification to the grammar corresponds to what is already implemented
by the compiler.