PSA - Update the description of empty action selector groups ...
... in response to discussion at the last P4 architecture work group meeting.

Recommend, but do not require, that PSA data plane support empty
groups by executing the table's default action if an empty group is
encountered during packet processing time, with no restrictions on the
names of the default action vs. the actions used in the non-empty
groups.

Describe some details of a faithful empty group implementation in the
new appendix, and why the proposed workaround does not quite live up
to that standard.