PSA - Change prohibition to "need not be supported"
In general, if a feature is something we do not want to require PSA
implementations to support in some version of the PSA spec, it is
preferable to say "X need not be supported" rather than "you cannot do
X".  The former means a PSA implementation _may_ support it if they
want, and they should be able to.  Saying "you cannot do X" makes it
sound like an implementation will violate the PSA spec if it supports
features that are not required.