P4_16: Move text on reading uninit vals and writing to invalid headers
I moved this text into its own section, and added cross-references to
it from where the text was moved from, Section 8.14 "Operations on
headers", and another in Section 8.13 "Operations on structs", even
though it really does apply to reading uninitialized of _any_ types.

Motivation for this movement: The moved text specifially describes the
behavior when an uninitialized value of type `enum` or `error` is
read, even though a header is not allowed to contain fields with those
types.  With the text in Section 8.14, one could be easily be confused
into believing that it implied that a header could contain fields of
those types.

The only changes I made to the moved text are as follows:

+ Changed the first sentence to "The result of reading a field" to
  "The result of reading a value", since reading uninitialized values
  applies not only to fields of headers and structs, but also to local
  variables with types like `bit<W>` or `bool`.

+ Added a few more explicit examples of situations where an
  uninitialized value can be read, changing this text: "reading any
  other value that has not been initialized, e.g. a field from a
  `struct`." to this: "reading any other value that has not been
  initialized, e.g. a field from a `struct`, any uninitialized
  variable inside of an `action` or `control`, or an `out` parameter
  of a `control` or `action` you have called, which was not assigned a
  value during the execution of that `control` or `action` (this list
  of examples is not intended to be exhaustive)."

+ Added this paragraph at the end of the new section:
  "For a top level `parser` or `control` in an architecture, it is up
  to that architecture to specify whether `control` parameters with
  direction `in` or `inout` are initialized when the control is
  called, and under what conditions they are initialized, and if so,
  what their values will be."