Clarify the significance of order multiple table entries in 'const entries'
With this PR: https://github.com/p4lang/p4-spec/pull/737 the PSA
architecture now defines the significant of the order of multiple
entries in tables with a 'const entries' table property for all kinds
of tables, in a way that is consistent with the changes proposed here.

I will soon create a similar PR for the v1model architecture for BMv2
simple_switch that makes the same kinds of statements for it.