Mention that an enum with an underlying type is also called serializable
Before this change, the word "serializable" does not occur anywhere in
the language specification.  An enum value with an underlying type is
often called a serializable enum in conversation between P4
implementers, including in the p4c implementation, so having a way to
find it in the language spec by that name seems like a good idea.