Changes to allowed/error for two entries of proposed new table
Change `typedef int X` from error to allowed.  p4c now allows this
with recent changes related to making `int` a type usable in a P4
program.

Change `type tuple ... X` from allowed to error.  It was only allowed
in the first version of this table because p4c allowed it, but on
further review there is no known use case that needs such a
capability, so it seems best to keep the allowed list shorter for now,
and we can expand it later on an as-needed basis.