Proposed new error value for implementations that restrict parser alignment
Similar to the existing statement that some implemtations are allowed
to restrict headers to be multiples of 8 bits long, such
implementations are also likely to impose similar restrictions on the
2-argument extract and advance calls, so that the parsing 'cursor' is
always at a multiple of 8 bits from the beginning of the packet.
These changes propose a way to make that explicit in the
specification.