Add proposed figure showing PSA with all packer & unpacker controls
I would not propose _only_ showing this figure, as it is a lot of
detail to handle when you first look at this.  The existing similar
block diagram in the PSA spec is the first thing people should see,
since it emphasizes the parsers and controls that they _must_ write
code for.  This one shows all of the other controls that they _might_
want to write additional code for, if they want to preserve
user-defined metadata on one or more of the various packet paths.