Add a DigestCreator control for creation of digest messages
It is also run just after ingress, just like several of the new Packer
controls.  The intent is that now the IngressDeparser control will do
nothing but emit calls on headers, header stacks, and header_union
values.  Digest creation, which was formerly restricted to be done
within the IngressDeparser, would now be restricted to occur only in
the DigestCreator control.  It has a default EmptyDigestCreator
implementation that never generates any digest messages.